Allow to overwrite implementation details based on the request.connection
This commit introduces three new methods to the options, providing the developer using the middleware with the ability to modify how the NTLM user data and proxy cache ID are stored. This enhancement enables the middleware to be used with a (reverse) proxy if the application can supply a unique identifier for the user.

The default behavior of storing the user data and proxy cache ID on the request.connection is maintained as the default value for the new options.